se agrego las validaciones a cada endpoint

+++ routers/todos.py
@@ -29,12 +29,22 @@ class TodoRequest(BaseModel):
 # read_all() usa esa conexión para devolver todos los registros.
 @router.get("/", status_code=status.HTTP_200_OK)
 async def read_all(user: user_dependency, db: db_dependency):
+
+    if user is None:
+
+        raise HTTPException(status_code=401, detail='Authentication Fail')
+
     return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 #Get only ID
 @router.get("/todo/{todo_id}", status_code= status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id:int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Fail')
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id).
+                  filter(Todos.owner_id == user.get('id')).first())
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo not found!!!.')
@@ -51,8 +61,13 @@ async def create_todo(user:user_dependency, db: db_dependency, todo_request: Tod
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency, todo_request:TodoRequest, todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user:user_dependency, db:db_dependency, todo_request:TodoRequest, todo_id:int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Fail')
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id')).first())
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
@@ -67,13 +82,17 @@ async def update_todo(db:db_dependency, todo_request:TodoRequest, todo_id:int = 
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db:db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id')).first())
+
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Fail')
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
